requirements.txt
PyYAML==3.10
nltk==2.0.4
numpy==1.7.1
pyenchant==1.6.5
pymongo==2.4.2
wsgiref==0.1.2
